exa: fix drm/loongson + etnaviv usage
we want etnaviv export buffer, and xf86-video-loongson and mesa
sharing the buffer, they are two different process , but this is
self sharing. once xf86-video-loongson get the sharing dmabuf , it
can prime to handle. then cpu access is possible. let the cpu do the
resolve work. resolve or copy , depending on the tiling of the buffer
object exported.

Signed-off-by: suijingfeng <suijingfeng@loognson.cn>